requirements.txt
cudatoolkit=10.2.89
numpy=1.19.1
pandas=1.1.0
python=3.8.5
scikit-learn=0.23.2
pytorch=1.6.0